% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{culturemapdata}
\alias{culturemapdata}
\title{culturemapdata}
\format{
A data frame with 131 rows and 4 variables:
\describe{
\item{\code{HRAF_ID}}{character HRAF OWC Culture ID.}
\item{\code{Subsistence Type}}{character HRAF subsistence type for society.}
\item{\code{latitude}}{double Latitude for society location.}
\item{\code{longitude}}{double Longitude for society location.}
}
}
\usage{
culturemapdata
}
\description{
A data frame for producing a map of societies in the sample.
}
\keyword{datasets}
