% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{documentdata}
\alias{documentdata}
\title{document data}
\format{
A data frame with 131 rows and 6 variables:
\describe{
\item{\code{EHRAF_NAME}}{character eHRAF culture name.}
\item{\code{document_ID}}{character eHRAF document ID for document.}
\item{\code{culture_ID}}{double eHRAF OWC ID for culture.}
\item{\code{document_publication_date}}{integer Document publication year.}
\item{\code{document_page_count}}{integer Documentn page count.}
\item{\code{female_coauthor_present}}{integer Presence of a female author or co-author. Coded as 1 for present, 0 for absent.}
}
}
\usage{
documentdata
}
\description{
A data frame of document-level metadata.
}
\keyword{datasets}
