% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{punishments_data_long}
\alias{punishments_data_long}
\title{punishments_data_long}
\format{
A data frame with 2620 rows and 5 variables:
\describe{
\item{\code{HRAF_ID}}{character HRAF OWC Culture ID.}
\item{\code{Coding}}{double Coding for culture by variable. Coded as 1 for evidence for, 0 for no evidence.}
\item{\code{Domain}}{character Domain of norm violation type being coded.}
\item{\code{Sanction}}{character Punishment type being coded.}
\item{\code{Coding_label}}{character Coding for culture by varable, with text as label.}
}
}
\usage{
punishments_data_long
}
\description{
A long-form version of the data of norm violation, punishments, and codings for all cultures.
}
\keyword{datasets}
